MediaManager: Get ready for rewrite
The MediaManager has a major flaw that has been found many months before. However, the flaw was not an issue that would undermine the basic implementation of MaterialTunes until now. The MediaPlayer
object within the MediaManager Class follows a music playlist in accordance with the songFiles ArrayList. This is considered dangerous as there won't be a proper way of managing album or playlist
specific media playing and we would have to perform checks everytime we play an entire album. Instead, we'll create a new queue for the MediaManager so that the MediaPlayer Object can rely on the
queue to perform media playback without any form of song metadata checks.

Signed-off-by: nixxholas <nixholas@outlook.com>