Source: Queue Update
This commit will address the Queue issue with the MediaManager Object and will help to simplify the process of adding an album to queue, or adding a playlist to queue, etc.
Future commits will implement such features to ensure that the Music Player is flexible in complex environments or algorithms towards media playback management.

Signed-off-by: nixxholas <nixholas@outlook.com>