SongAdapter: currentlyPlayingIndex Deadlock Fix
This commit addresses the issue whereby the app would crash upon playing a song as the currentlyPlayingIndex is set via the songFiles list, which is substantially bigger than the queue. This would
in turn cause an overflow. We replace it by directly linking currentlyPlayingIndex to the queue instead of the songFiles list.

Signed-off-by: nixxholas <nixholas@outlook.com>