Source: Integer Based Queue & Upcoming System
The old method of using song objects results in a very low tolerance in memory integrity resulting in quicker crashes. Using integers only not only helps reduce memory usage, but at the same time allows us to have a large queue stack.

We are left with the notifications to resolve.

Signed-off-by: nixxholas <nixholas@outlook.com>